add chose origin

+++ sql-script.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS schedule(
     destination VARCHAR(20) NOT NULL,
     departure DATETIME NOT NULL,
     distance DOUBLE NOT NULL,
-    price DECIMAL(5, 2) NOT NULL # 3 digits before the decimal, 2 after
+    price FLOAT NOT NULL # 3 digits before the decimal, 2 after
 ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=latin1;
 
 INSERT INTO schedule(destination, departure, distance, price) VALUES
